Move tests from `myft/test/ui` to `test`
 🐿 v2.10.2
