+++ karma.conf.js
@@ -35,14 +35,14 @@ module.exports = function (karma) {
 				'Array.prototype.@@iterator'
 			].join(','),
 			'components/unread-articles-indicator/test/*.spec.js',
-			'myft/**/*.spec.js'
+			'test/**/*.spec.js'
 		],
 
 		// preprocess matching files before serving them to	the browser
 		// available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
 		preprocessors: {
 			'components/unread-articles-indicator/test/*.spec.js': ['webpack', 'sourcemap'],
-			'myft/**/*.spec.js': ['webpack', 'sourcemap']
+			'test/**/*.spec.js': ['webpack', 'sourcemap']
 		},
 		webpack: {
 			module: {
