--- test/main.js
@@ -1,2 +1,0 @@
-require('../myft');
-require('../myft-common');
