--- test/main.scss
@@ -1,1 +1,0 @@
-@import '../myft/main';
