+++ webpack.config.js
@@ -6,8 +6,11 @@ module.exports = nWebpack({
 	withBabelPolyfills: false,
 	withHeadCss: false,
 	entry: {
-		'./public/main-without-n-ui.js': './test/main.js',
-		'./public/main.css': './test/main.scss'
+		'./public/main-without-n-ui.js': [
+			'./myft/index.js',
+			'./myft-common/index.js'
+		],
+		'./public/main.css': './myft/main.scss'
 	},
 	includes: [
 		__dirname
